Added listings with airbnb api call

+++ db/migrate/20201105075700_devise_create_users.rb
@@ -8,7 +8,7 @@ class DeviseCreateUsers < ActiveRecord::Migration[6.0]
       t.string :encrypted_password, null: false, default: ""
       t.string :first_name
       t.string :last_name
-      t.integer :code
+      t.bigint :code
 
       ## Recoverable
       t.string   :reset_password_token
